Seperate Invite and Activity pane
