+++ client/web/src/auth/welcome/InviteCollaborators/InviteCollaborators.module.scss
